docs: Make setup instructions use 'latest' branch instead of 'master'.
Because from experience we are seeing that the `master` branch is not
always deployable, we need some sort of light-weight release process.

Once every two weeks, we will tag a version, manually test it, and update
the `latest` branch. This will then also trigger an update of the Docker
image.